Rewrite skills code
Fixes #1091

+++ src/vocation.h
@@ -34,7 +34,7 @@ class Vocation
 		const std::string& getVocDescription() const {
 			return description;
 		}
-		uint32_t getReqSkillTries(int32_t skill, int32_t level);
+		uint64_t getReqSkillTries(uint8_t skill, uint16_t level);
 		uint64_t getReqMana(uint32_t magLevel);
 
 		uint16_t getId() const {
